// NTG-Admin-main/Front/ntg_admin/src/app/Models/engineer.ts
export interface Engineer {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  address: string;
  education: string;
  employmentHistory: string;
  numberOfYearsOfExperience: number | null;
  birthDate: Date;
  gender: string;
  religion: string;
  nationalNumber: number;
  status: boolean;
}